+++ CHANGELOG.md
@@ -1,2 +1,8 @@
+## 0.0.2
+
+- Added `HStack` widget.
+- Added `VStack` widget.
+- Added `Rectangle` widget.
+
 ## 0.0.1 - Sept, 2023
 Setting up the project structure. This release is not usable.